continue work on refactoring transactions
